@@ -2,6 +2,7 @@
   "name": "docu-notion-sample-site",
   "version": "0.0.0",
   "private": false,
+  "type": "module",
   "scripts": {
     "clear-docs": "rimraf ./docs/",
     "pull-test-tagged": "yarn clear-docs && cross-var yarn docu-notion -n $DOCU_NOTION_INTEGRATION_TOKEN -r $DOCU_NOTION_SAMPLE_ROOT_PAGE -m ./docs -i ./static/notion_imgs -p /notion_imgs/ --status-tag test",
@@ -15,23 +16,24 @@
     "serve": "docusaurus serve"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.4.3",
-    "@docusaurus/plugin-google-gtag": "^3.5.2",
-    "@docusaurus/preset-classic": "^2.4.3",
-    "@easyops-cn/docusaurus-search-local": "^0.36.0",
-    "@mdx-js/react": "^1",
+    "@docusaurus/core": "^3.7.0",
+    "@docusaurus/plugin-google-gtag": "^3.7.0",
+    "@docusaurus/preset-classic": "^3.7.0",
+    "@easyops-cn/docusaurus-search-local": "^0.40.0",
+    "@mdx-js/react": "^3.0.0",
     "@sillsdev/docu-notion": "^0.16.1",
-    "clsx": "^1.2.1",
-    "prism-react-renderer": "^1.3.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "clsx": "^2.0.0",
+    "prism-react-renderer": "^2.4.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-gif-player": "^0.4.2",
     "react-player": "^2.10.1",
     "rimraf": "^4.1.2"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^2.4.3",
-    "@tsconfig/docusaurus": "^2.0.1",
+    "@docusaurus/module-type-aliases": "^3.7.0",
+    "@docusaurus/tsconfig": "^3.7.0",
+    "@types/node": "^20.11.0",
     "cross-var": "^1.1.0",
     "typescript": "^5.2.2"
   },
@@ -53,7 +55,7 @@
     "yarn": "1.22.19"
   },
   "overrides": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   }
 }
