@@ -1,3 +1,3 @@
-module.exports = {
-  presets: [require.resolve('@docusaurus/core/lib/babel/preset')],
+export default {
+  presets: ['@docusaurus/core/lib/babel/preset'],
 };
