@@ -1,8 +1,8 @@
 // @ts-check
 // Note: type annotations allow type checking and IDEs autocompletion
 
-const lightCodeTheme = require("prism-react-renderer/themes/github");
-const darkCodeTheme = require("prism-react-renderer/themes/dracula");
+import {themes as prismThemes} from 'prism-react-renderer';
+import path from 'path';
 
 /** @type {import('@docusaurus/types').Config} */
 const config = {
@@ -25,13 +25,13 @@ const config = {
       ({
         docs: {
           routeBasePath: "/", // Serve the docs at the site's root
-          sidebarPath: require.resolve('./sidebars.js'), //  `undefined` to create a fully autogenerated 
+          sidebarPath: path.resolve('./sidebars.js'), //  `undefined` to create a fully autogenerated 
         }, 
         blog: false,
         theme: {
           customCss: [
-            require.resolve("./css/docu-notion-styles.css"),
-            require.resolve(
+            path.resolve("./css/docu-notion-styles.css"),
+            path.resolve(
               "./css/gifplayer.css"
               //"./node_modules/react-gif-player/dist/gifplayer.css" // this gave a big red compile warning which is seaming unrelated "  Replace Autoprefixer browsers option to Browserslist config..."
             ),
@@ -59,7 +59,7 @@ const config = {
   themes: [
     // ... Your other themes.
     [
-      require.resolve("@easyops-cn/docusaurus-search-local"),
+      '@easyops-cn/docusaurus-search-local',
       {
         // ... Your options.
         // `hashed` is recommended as long-term-cache of index file is possible.
@@ -156,10 +156,10 @@ const config = {
       }, 
 
       prism: {
-        theme: lightCodeTheme,
-        darkTheme: darkCodeTheme,
+        theme: prismThemes.github,
+        darkTheme: prismThemes.dracula,
       },
     }),
 };
 
-module.exports = config;
+export default config;
