@@ -166,4 +166,4 @@ const sidebars = {
    
 };
 
-module.exports = sidebars;
+export default sidebars;
